+++ tests/bdd/watch_loop/watch_loop_test.go
@@ -1,0 +1,21 @@
+package watch_loop
+
+import (
+	"testing"
+
+	"github.com/cucumber/godog"
+	"github.com/jorelcb/codify/tests/bdd/commons"
+)
+
+func TestWatchLoopFeature(t *testing.T) {
+	suite := godog.TestSuite{
+		Name:                 "watch_loop",
+		TestSuiteInitializer: InitializeTestSuite,
+		ScenarioInitializer:  InitializeScenario,
+		Options:              commons.Options("./watch_loop.feature"),
+	}
+
+	if suite.Run() != 0 {
+		t.Fatal("non-zero status returned, failed to run feature tests")
+	}
+}